rm error

+++ app/controllers/books_controller.rb
@@ -23,15 +23,22 @@ class BooksController < ApplicationController
   end
 
   def update
-    book = Book.find(params[:id])
-    book.update(book_params)
-    redirect_to book_path(book), notice: 'Book was successfully updated.'
+  	@book = Book.find(params[:id])
+    if @book.update(book_params)
+      redirect_to book_path(@book.id), notice: 'Book was successfully updated.'
+    else
+      render 'edit'
+    end
   end
 
   def destroy
-    book = Book.find(params[:id])
-    book.destroy
-    redirect_to books_path, notice: 'Book was successfully destroyed.'
+    if book = Book.find(params[:id])
+       book.destroy
+       redirect_to books_path, notice: 'Book was successfully destroyed.'
+    else
+      @books = Book.all
+      render '_form'
+    end
   end
 
   private
